more formatting updates
